fix clang warnings
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@1979 2e953b5c-c64d-0410-be54-f773e93e544c

+++ smartgame/SgFastLog.cpp
@@ -34,10 +34,10 @@ SgFastLog::SgFastLog(int mantissaBits)
     x.m_int = 0x3F800000;
     int incr = (1 << m_mantissaBitsDiff);
     int p = static_cast<int>(pow(2.0f, mantissaBits));
-    float invLogTwo = 1.f / log(2.f);
+    float invLogTwo = 1.f / logf(2.f);
     for (int i = 0; i < p; ++i)
     {
-        m_lookupTable[i] = log(x.m_float) * invLogTwo;
+        m_lookupTable[i] = logf(x.m_float) * invLogTwo;
         x.m_int += incr;
     }
 }
